requirements.txt
pyspark==3.3.0
delta-spark==2.2.0